Separate integration and unit tests
